requirements.txt
pandas==2.1.0
python-docx==0.8.11
openpyxl==3.1.2
pyinstaller==5.13.2